+++ src/main/kotlin/com/tul/shoppingcart/exceptions/BadRequestException.kt
@@ -1,0 +1,3 @@
+package com.tul.shoppingcart.exceptions
+
+class BadRequestException(val errorCode: ErrorCode): Exception(errorCode.msg)