+++ src/main/kotlin/com/tul/shoppingcart/services/ProductService.kt
@@ -1,7 +1,10 @@
 package com.tul.shoppingcart.services
 
+import com.tul.shoppingcart.converters.RequestConverter
+import com.tul.shoppingcart.dto.ProductDTO
 import com.tul.shoppingcart.repository.ProductRepository
 import com.tul.shoppingcart.repository.model.Product
+import com.tul.shoppingcart.validations.RequestValidator
 import org.springframework.stereotype.Service
 
 @Service
@@ -11,5 +14,10 @@ class ProductService(
 
     fun getAllProducts(): MutableIterable<Product> {
         return productRepository.findAll()
+    }
+
+    fun saveProduct(request: ProductDTO): String {
+        RequestValidator.validateProductToSave(request)
+        return productRepository.save(RequestConverter.convert(request)).id.toString()
     }
 }