+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 
 This file will show the step by step of the commits since project's creation
 
+## 17/01/2022 17:30 - Branch: feature/services
+- Validations and global exception handler
+![img_1.png](img_1.png)
+
 ## 17/01/2022 15:40 - Branch: feathre/h2-database
 - H2 database to use as a repository of products and carts.